Extract credentials passed in Url
This commit extracts the credentials passed in a Url and
uses them to pass Basic Credentials when constructing
the single node Transport, as well as constructing a client
for elasticsearch integration tests.

Leaving the credentials in the Url and
also configuring credentials through .auth() fn, results
in reqwest sending two Authorization headers.